Move common React components to separate directory
