+++ extensions/ql-vscode/src/view/common/HorizontalSpace.tsx
@@ -1,9 +1,7 @@
 import styled from 'styled-components';
 
-const HorizontalSpace = styled.div<{ size: 1 | 2 | 3 }>`
+export const HorizontalSpace = styled.div<{ size: 1 | 2 | 3 }>`
   flex: 0 0 auto;
   display: inline-block;
   width: ${props => 0.2 * props.size}em;
 `;
-
-export default HorizontalSpace;
