+++ extensions/ql-vscode/src/view/common/SectionTitle.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const SectionTitle = styled.h2`
+export const SectionTitle = styled.h2`
   font-size: medium;
   font-weight: 500;
   padding: 0 0.5em 0 0;
@@ -8,5 +8,3 @@ const SectionTitle = styled.h2`
   display: inline-block;
   vertical-align: middle;
 `;
-
-export default SectionTitle;
