+++ extensions/ql-vscode/src/view/common/ViewTitle.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const ViewTitle = styled.h1`
+export const ViewTitle = styled.h1`
   font-size: 2em;
   margin-bottom: 0.5em;
   font-weight: 500;
@@ -9,5 +9,3 @@ const ViewTitle = styled.h1`
   overflow: hidden;
   text-overflow: ellipsis;
 `;
-
-export default ViewTitle;
