+++ extensions/ql-vscode/src/view/common/index.ts
@@ -1,0 +1,4 @@
+export * from './HorizontalSpace';
+export * from './SectionTitle';
+export * from './VerticalSpace';
+export * from './ViewTitle';
