+++ extensions/ql-vscode/src/view/variant-analysis/QueryDetails.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import ViewTitle from '../remote-queries/ViewTitle';
+import { ViewTitle } from '../common';
 import { LinkIconButton } from './LinkIconButton';
 
 export type QueryDetailsProps = {
